\name{crab}
\alias{crab}
\docType{data}
\title{PSR J0534+2200 (Crab-Pulsar) Time of Arrivals}
\description{
This data set contains n=21145 time of arrivals of photons with energies above 100MeV of PSR J0534+2200 (Crab-pulsar), obtained from the Fermi LAT.}
\usage{data(crab)}
\format{A vector containing 21145 observation.   
}

\source{Obtained from Fermi LAT, energies above 100 MeV.}
\references{
Abdo A, et al. (2010b). Fermi large area telescope observations of the Crab pulsar and
nebula. The Astronomical Journal, 708, 1254-1267. \cr
}
\examples{
data(crab)
SOPIE(crab)
}
\keyword{datasets}
