\name{J1709}
\alias{J1709}
\docType{data}
\title{PSR J1709-44290 Time of Arrivals}
\description{
This data set contains n=21153 time of arrivals of photons with energies above 100MeV of PSR J1709-44290, obtained from the Fermi LAT.}
\usage{data(J1709)}
\format{A vector containing 21153 observation.   
}
\source{Obtained from Fermi LAT, energies above 100 MeV}
\references{
Abdo A, et al. (2010). "The first Fermi large area telescope catalog of gamma-ray pulsars."
The Astrophysical Journal Supplement Series, 187, 460-494. \cr
}
\examples{
data(J1709)
SOPIE(J1709)
}
\keyword{datasets}
